// app/src/models/Event/Event.ts
export interface EventInterface {
  id: string;
  name: string;
  eventKey?: string;
  description: string;
  cost: number;
  currency: "₹";
  media: (CloudinaryMedia | AWSMedia)[];
  thumbnail?: CloudinaryMedia | AWSMedia;

  freeGame?: boolean;

  ownerUID: string;
  createdOn: number;
  updatedOn: number;
  // status: "live" | "draft";
  joinURL: string;

  earnings: number;
  views: number;
  students: number;
  acceptInvites?: boolean;
  // eventDateTime?: string;
  soldOut?: boolean;
  eventDateTimeList?: string[];
  programDetails?: ListItem[]; // program details
  whoIsItFor?: ListItem[]; // what you get
  faq?: ListItem[];
  courseGoal?: string;
  courseGoalPrimary?: string;

  reminderTemplate?: "event_update";
  googleTitle?: string;
  googleDescription?: string;
  googleSEOImg?: CloudinaryMedia | AWSMedia;

  eventType?: eventTypes;

  program?: SessionV2[];
  communityVisibility?: boolean;

  needsRegistration?: boolean;
  orderNumber?: number;
  parentId?: string;
  ownerDetails?: {
    name?: string;
    img?: CloudinaryMedia;
    userKey?: string;
  };
  celebrityId?: string;

  state?: "active" | "finished";
  eventStarts?: number; // start time unix

  goalPoints?: string[];

  calCritieron?: number;
  nbTotalParticipants?: number;
  // engineChoice?: engineChoice;
  challengeLength?: number;
  calThForStreak?: number;
  streakLengthTh?: number;

  leaderboards?: LeaderboardDescription[];

  seriesAccess?: "GLOBAL";

  recentActivity?: UserActivitySnippet;
  challengeType?: "game";
  awardsWorth?: number;
  enrolledUserUIDs?: string[];

  manualRank?: number;
  sprintLength?: number;
  roundLength?: number;

  configuration?: GameConfiguration;

  badges?: Badge[];

  pricing?: PricingPlan[];
  nbWorkouts?: number;
  equipmentNeed?: string;
  goalKpiText?: string;

  inviteCode?: string;
}

export interface PricingPlan {
  id: string;
  name: string;
  cost: number;
  currency: string;
  appliedOn: "round" | "sprint";
  moneyBackDays?: number;
  freeAccessDays?: number;
  durationInDays?: number;
  pointers?: string[];
}

export interface UserActivitySnippet {
  uid: string;
  createdOn: number;
  postId: string;
  text: string;
  createdBy?: string;
  img?: CloudinaryMedia;
}

export interface LeaderboardDescription {
  name: string;
  kpi: "streak" | "calories" | "avgSpeed" | "distance";
}

export type leaderboardKPIs = "streak" | "calories" | "avgSpeed" | "distance";

interface SessionKPIObj {
  [sessionId: string]: {
    numClaps?: number;
    numCheckins?: number;
  };
}

export interface EventKPIs {
  sessionKPIs: SessionKPIObj;
}

export type eventTypes = "challenge" | "course";
// ...
export type sessionTypes = "live" | "ondemand" | "activity" | "post";

export interface SessionV2 {
  id: string;
  sessionType: sessionTypes;
  name?: string;
  description?: string;
  media?: CloudinaryMedia;
  youtubeLink?: string;

  dayNumber?: number;

  joinURL?: string;
  live?: boolean;
  free?: boolean;

  updatedOn: number;

  numClaps?: number;
  numCheckins?: number;
  sessionStarts?: string;

  cohortId?: string;
}
// ...
export type gameTypes = "team" | "individual";
export interface GameConfiguration {
  rounds: RoundObject[];
  starts: number; // unix
  challengeLength: number;

  sprints: SprintObject[];
  kpis?: GameKPITarget[];
  topNav?: navItemsType[];

  stepTaskId?: string;

  // payment collection for sprint
  activeSprintId?: string;
  pinnedSprintId?: string;

  gameType?: gameTypes;
  tagTypes?: string[];
  popularSearches?: string[];
}

export interface GameKPITarget {
  kpi: SystemKPIs;
  targetValue: number;
}

// pullups - pullups, reps, targetVal

export interface RoundObject {
  length: number;
  numPrizes: number;
  name: string;
  id: string;

  type: "REST" | "ACTIVE";
  sprintId: string;
  isFinale?: boolean;
  isWarmup?: boolean;
}

export interface SprintObject {
  length: number;
  numPrizes: number;
  name: string;
  // description?: string;

  id: string;

  startString?: string;
  endString?: string;
  filter?: "gender:female";
  // registrationExpires?: number;
}